+++ src/main/kotlin/com/github/jensim/megamanipulator/settings/types/codehost/GitHubSettings.kt
@@ -2,14 +2,13 @@ package com.github.jensim.megamanipulator.settings.types.codehost
 
 import com.fasterxml.jackson.annotation.JsonPropertyDescription
 import com.github.jensim.megamanipulator.settings.types.AuthMethod
-import com.github.jensim.megamanipulator.settings.types.AuthMethod.USERNAME_TOKEN
+import com.github.jensim.megamanipulator.settings.types.AuthMethod.JUST_TOKEN
 import com.github.jensim.megamanipulator.settings.types.CloneType
 import com.github.jensim.megamanipulator.settings.types.CloneType.SSH
 import com.github.jensim.megamanipulator.settings.types.ForkSetting
 import com.github.jensim.megamanipulator.settings.types.ForkSetting.LAZY_FORK
 import com.github.jensim.megamanipulator.settings.types.HttpsOverride
 import com.github.jensim.megamanipulator.settings.types.codehost.CodeHostSettingsType.GITHUB
-import com.github.jensim.megamanipulator.settings.types.encodeToBase64String
 import com.github.jensim.megamanipulator.settings.types.forkSettingDescription
 
 data class GitHubSettings(
@@ -24,7 +23,7 @@ data class GitHubSettings(
 ) : CodeHostSettings() {
 
     override val codeHostType: CodeHostSettingsType = GITHUB
-    override val authMethod: AuthMethod = USERNAME_TOKEN
+    override val authMethod: AuthMethod = JUST_TOKEN
     override val baseUrl: String = "https://api.github.com"
     val graphQLUrl: String = "https://graphql.github.com/graphql/proxy"
 
@@ -34,7 +33,7 @@ data class GitHubSettings(
 
     override fun getAuthHeaderValue(password: String?): String? = when {
         // https://docs.github.com/en/rest/guides/getting-started-with-the-rest-api#authentication
-        password != null && authMethod == USERNAME_TOKEN -> "Basic ${encodeToBase64String("$username:$password")}"
+        password != null && authMethod == JUST_TOKEN -> "token $password"
         else -> null
     }
 }