+++ src/test/kotlin/com/github/jensim/megamanipulator/settings/types/HttpLoggingLevelTest.kt
@@ -1,0 +1,16 @@
+package com.github.jensim.megamanipulator.settings.types
+
+import io.ktor.client.features.logging.LogLevel
+import org.hamcrest.MatcherAssert.assertThat
+import org.hamcrest.Matchers.equalTo
+import org.junit.jupiter.api.Test
+
+internal class HttpLoggingLevelTest {
+    @Test
+    internal fun `all match to ktor`() {
+        assertThat(
+            HttpLoggingLevel.values().map { it.name }.sorted(),
+            equalTo(LogLevel.values().map { it.name }.sorted())
+        )
+    }
+}